Misc tweaks
Fix ug editor again
https://github.com/kwsch/PKHeX/commit/b4aeb1ac7c989d410ec581cfb107861b8ffcb0ba

make framegenerator info a struct (8 bytes); will be rewritten later anyway

+++ PKHeX.Core/Legality/RNG/Frame/FrameFinder.cs
@@ -21,7 +21,7 @@ public static class FrameFinder
             return [];
 
         // Don't trust pk.Nature, just get the correct original via EncryptionConstant
-        var info = new FrameGenerator(pk) {Nature = pk.EncryptionConstant % 25};
+        var info = new FrameGenerator(pk) {Nature = (byte)(pk.EncryptionConstant % 25)};
         var seeds = GetSeeds(pidiv, info, pk);
         var frames = pidiv.Type == PIDType.CuteCharm
             ? FilterCuteCharm(seeds, info)
